k8scluster: break out SyncMetadata tests
