+++ receiver/k8sclusterreceiver/collection/pods_test.go
@@ -30,7 +30,6 @@ import (
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/common/testing/util"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/metadata"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/k8sclusterreceiver/testutils"
 )
@@ -206,145 +205,6 @@ func TestPhaseToInt(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := phaseToInt(tt.phase); got != tt.want {
 				t.Errorf("phaseToInt() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-var commonPodMetadata = map[string]string{
-	"foo":                    "bar",
-	"foo1":                   "",
-	"pod.creation_timestamp": "0001-01-01T00:00:00Z",
-}
-
-var allPodMetadata = func(metadata map[string]string) map[string]string {
-	out := util.MergeStringMaps(metadata, commonPodMetadata)
-	return out
-}
-
-func TestDataCollectorSyncMetadata(t *testing.T) {
-	tests := []struct {
-		name          string
-		metadataStore *metadataStore
-		resource      interface{}
-		want          map[metadata.ResourceID]*KubernetesMetadata
-	}{
-		{
-			name:          "Pod and container metadata simple case",
-			metadataStore: &metadataStore{},
-			resource: newPodWithContainer(
-				"0",
-				podSpecWithContainer("container-name"),
-				podStatusWithContainer("container-name", "container-id"),
-			),
-			want: map[metadata.ResourceID]*KubernetesMetadata{
-				metadata.ResourceID("test-pod-0-uid"): {
-					resourceIDKey: "k8s.pod.uid",
-					resourceID:    "test-pod-0-uid",
-					metadata:      commonPodMetadata,
-				},
-				metadata.ResourceID("container-id"): {
-					resourceIDKey: "container.id",
-					resourceID:    "container-id",
-					metadata: map[string]string{
-						"container.status": "running",
-					},
-				},
-			},
-		},
-		{
-			name:          "Empty container id skips container resource",
-			metadataStore: &metadataStore{},
-			resource: newPodWithContainer(
-				"0",
-				podSpecWithContainer("container-name"),
-				podStatusWithContainer("container-name", ""),
-			),
-			want: map[metadata.ResourceID]*KubernetesMetadata{
-				metadata.ResourceID("test-pod-0-uid"): {
-					resourceIDKey: "k8s.pod.uid",
-					resourceID:    "test-pod-0-uid",
-					metadata:      commonPodMetadata,
-				},
-			},
-		},
-		{
-			name:          "Pod with Owner Reference",
-			metadataStore: &metadataStore{},
-			resource: withOwnerReferences([]v1.OwnerReference{
-				{
-					Kind: "StatefulSet",
-					Name: "test-statefulset-0",
-					UID:  "test-statefulset-0-uid",
-				},
-			}, newPodWithContainer("0", &corev1.PodSpec{}, &corev1.PodStatus{})),
-			want: map[metadata.ResourceID]*KubernetesMetadata{
-				metadata.ResourceID("test-pod-0-uid"): {
-					resourceIDKey: "k8s.pod.uid",
-					resourceID:    "test-pod-0-uid",
-					metadata: allPodMetadata(map[string]string{
-						"k8s.workload.kind": "StatefulSet",
-						"k8s.workload.name": "test-statefulset-0",
-						"statefulset":       "test-statefulset-0",
-						"statefulset_uid":   "test-statefulset-0-uid",
-					}),
-				},
-			},
-		},
-		{
-			name: "Pod with Service metadata",
-			metadataStore: &metadataStore{
-				services: &testutils.MockStore{
-					Cache: map[string]interface{}{
-						"test-namespace/test-service": &corev1.Service{
-							ObjectMeta: v1.ObjectMeta{
-								Name:      "test-service",
-								Namespace: "test-namespace",
-								UID:       "test-service-uid",
-							},
-							Spec: corev1.ServiceSpec{
-								Selector: map[string]string{
-									"k8s-app": "my-app",
-								},
-							},
-						},
-					},
-				},
-			},
-			resource: podWithAdditionalLabels(
-				map[string]string{"k8s-app": "my-app"},
-				newPodWithContainer("0", &corev1.PodSpec{}, &corev1.PodStatus{}),
-			),
-			want: map[metadata.ResourceID]*KubernetesMetadata{
-				metadata.ResourceID("test-pod-0-uid"): {
-					resourceIDKey: "k8s.pod.uid",
-					resourceID:    "test-pod-0-uid",
-					metadata: allPodMetadata(map[string]string{
-						"kubernetes_service_test-service": "",
-						"k8s-app":                         "my-app",
-					}),
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		observedLogger, _ := observer.New(zapcore.WarnLevel)
-		logger := zap.New(observedLogger)
-		t.Run(tt.name, func(t *testing.T) {
-			dc := &DataCollector{
-				logger:                 logger,
-				metadataStore:          tt.metadataStore,
-				nodeConditionsToReport: []string{},
-			}
-
-			actual := dc.SyncMetadata(tt.resource)
-			require.Equal(t, len(tt.want), len(actual))
-
-			for key, item := range tt.want {
-				got, exists := actual[key]
-				require.True(t, exists)
-				require.Equal(t, *item, *got)
 			}
 		})
 	}
